requirements.txt
pandas
openpyxl
pyyaml
requests
Faker
dbt-snowflake
sqlfmt